requirements.txt
Django==2.2.3
django-bootstrap4==0.0.8
django-extensions==2.2.1
Pillow==6.1.0
pkg-resources==0.0.0
psycopg2==2.8.3
pydotplus==2.0.2
pyparsing==2.4.2
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
django-bootstrap-datepicker-plus

